chore(ui): split API key creation dialogs
